refactor(api): reduce Room to a thin DurableObject adapter over RoomCore
Phase 1 of specs/20260522T220000-api-runtime-portability.md, cut-over.

Moves and rewrites apps/api/src/room.ts as
apps/api/src/room/backends/cloudflare/durable-object.ts. The class is
now a thin shell: it constructs createDurableObjectUpdateLog(storage)
and createRoomCore({ updateLog }), and forwards every Hibernation API
callback (fetch, webSocketMessage, webSocketClose, webSocketError,
alarm) and RPC method (sync, getDoc) to the RoomCore instance.

Cloudflare's hibernation WebSocket structurally satisfies RoomSocket
(send/close/readyState), so the raw ws is passed straight to core
with no wrapper.

Other cut-over touches:
- registry.ts type-only import of Room now points at the moved file.
- app.ts re-export path updated so wrangler types resolves
  DurableObjectNamespace<Room>.
- room.test.ts also moved next to its subject as
  durable-object.test.ts; both dynamic imports updated to
  './durable-object.js'.
- app.rooms.test.ts swaps its Room-module mock for a cloudflare:workers
  mock. The route-boundary tests never construct Room, so loading the
  real class with a stubbed DurableObject base is harmless and avoids
  a mock that could leak into durable-object.test.ts when both files
  run in the same bun test process.

Verified: bun run typecheck clean; bun test reports 86 pass / 0 fail
across 11 files. wrangler dev continues to serve rooms unchanged.
